feat: VideoContainer 컴포넌트의 기능을 SearchResult와 VideoList 컴포넌트로 분리
